Refactor base view into seperate modules for easier maintenace and create seperate mixin module
