+++ netbox_librenms_plugin/views/sync/locations.py
@@ -8,7 +8,7 @@ from django_tables2 import SingleTableView
 
 from netbox_librenms_plugin.filtersets import SiteLocationFilterSet
 from netbox_librenms_plugin.tables import SiteLocationSyncTable
-from netbox_librenms_plugin.views.base_views import LibreNMSAPIMixin
+from netbox_librenms_plugin.views.mixins import LibreNMSAPIMixin
 
 
 class SyncSiteLocationView(LibreNMSAPIMixin, SingleTableView):
