+++ netbox_librenms_plugin/views/vm_views.py
@@ -5,8 +5,9 @@ from virtualization.models import VirtualMachine
 
 from netbox_librenms_plugin.tables import LibreNMSVMInterfaceTable
 
-from .base_views import (BaseInterfaceTableView, BaseIPAddressTableView,
-                         BaseLibreNMSSyncView)
+from .base.interfaces import BaseInterfaceTableView
+from .base.librenms_sync import BaseLibreNMSSyncView
+from .base.ip_addresses import BaseIPAddressTableView
 
 
 @register_model_view(VirtualMachine, name="librenms_sync", path="librenms-sync")
